requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
et-xmlfile==1.1.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
innovation-stk @ git+https://github.com/Fridaplatform/SofttekLLMSDK.git@4928d16fb90df48b32f1f251607b856b004669ee
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pandas-stubs==2.0.3.230814
Pillow==9.5.0
pinecone-client==2.2.4
protobuf==4.24.3
ptvsd==4.3.2
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
streamlit==1.27.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
types-pytz==2023.3.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.17.0
